Moved most of the details to Module_builder.md
This file had too much detail for a README, and much of it overlapped with what was described in docs/Module_builder.md. I moved most of the content to that file. I also added a section at the beginning describing prerequisites for those who want to try to use the SDK--Shane referred to this as the "you must be at least this tall to go on this ride" sign.

https://docs.google.com/document/d/1jMT6j_w_Oc3FiJDtgxiFaKHmWb9zXU0zT2s6ihckuE4/edit is the gdoc where Dylan and I commented on and edited the original contents of this README.