[WIP] use readlink to get paths at runtime
Please do not merge yet, untested!  It would be helpful for others to read and try out though.

The paths to scripts and libraries should be computed at runtime, since the make may be run on a different host than the container.  Use 

	script_dir=$(dirname "$(readlink -f "$0")")'

to get the cwd of the script.