Fix python server error handling, don't hardcode JAVA_HOME
Currently the server is dumping the __str__ representation of any
exceptions thrown by a server method (which includes the entire
stacktrace in the case of ServerErrors thrown by SDK clients) into the
message field of the error, which is meant for a end user readable
message. There's no way to present a reasonable error message to a user
with the current behavior, and neither the perl nor java servers behave
in the same way - they both provide short messages. Fix to copy the
message field of the caught exception as intended.

Also get rid of the seemingly unnecessary and highly annoying memoizing
of the JAVA_HOME envvar in the scripts.