Update py client to support dynamic service compilation.
TODO:
Do the same for the other clients.
Test for py3 compatibility. This is a big job; do it when servers are
also py3 compatible. Probably needs tests running in a docker container,
but this risks adding dependencies accidentally.
Test with dynamic services. Getting this running will be a nightmare.
Also probably requires a pre-configured docker container.