Fix tests & python build
Tests wouldn't pass due to wget missing certs for sourceforge.

make sdkbase failed for python builds, and kb_sdk test would also fail.
Update the installed security packages in sdkbase to the minimum to get
the build to work without resulting in SSL warnings all over the place
when the module is run.

Removed failing R tests as R will be deprecated in the next commit.